move uploadToWeb to the web exporter
